app/doormanager_test/requirements.txt
certifi
paho-mqtt
PyYAML
grpcio
grpcio-tools